Moving DijkstaNode in dijkstra_node.py
