+++ dijkstra_node.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+from dataclasses import dataclass
+
+
+@dataclass
+class DijkstraNode:
+    vertex: int
+    distance: float
+
+    def __lt__(self, other: DijkstraNode) -> bool:
+        return self.distance < other.distance
+
+    def __eq__(self, other: DijkstraNode) -> bool:
+        return self.distance == other.distance
